@@ -1,6 +1,6 @@
 {
   "name": "yummy.frontend",
-  "version": "0.6.0",
+  "version": "0.7.1",
   "author": "Yummy App - Place Side",
   "private": true,
   "scripts": {
@@ -14,24 +14,28 @@
     "start": "node server.js"
   },
   "dependencies": {
-    "@angular/common": "5.0.3",
-    "@angular/compiler": "5.0.3",
-    "@angular/compiler-cli": "5.0.3",
-    "@angular/core": "5.0.3",
-    "@angular/forms": "5.0.3",
-    "@angular/http": "5.0.3",
-    "@angular/platform-browser": "5.0.3",
-    "@angular/platform-browser-dynamic": "5.0.3",
+    "@angular/cli": "6.0.3",
+    "@angular/common": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "@angular/compiler-cli": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/forms": "6.0.3",
+    "@angular/http": "6.0.3",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
     "@ionic-native/camera": "4.3.3",
     "@ionic-native/core": "4.4.0",
     "@ionic-native/splash-screen": "4.4.0",
     "@ionic-native/status-bar": "4.4.0",
     "@ionic/storage": "2.1.3",
-    "@ngrx/effects": "^5.0.1",
-    "@ngrx/store": "^5.0.0",
-    "@ngx-translate/core": "8.0.0",
-    "@ngx-translate/http-loader": "^2.0.0",
-    "@stomp/ng2-stompjs": "4.0.0",
+    "@ngrx/core": "^1.2.0",
+    "@ngrx/effects": "^5.2.0",
+    "@ngrx/entity": "^6.1.0",
+    "@ngrx/store": "^5.2.0",
+    "@ngrx/store-devtools": "^6.1.0",
+    "@ngx-translate/core": "10.0.2",
+    "@ngx-translate/http-loader": "^4.0.0",
+    "@stomp/ng2-stompjs": "^6.0.1",
     "@stomp/stompjs": "^4.0.7",
     "@types/sockjs-client": "^1.1.0",
     "cordova-ios": "~4.5.4",
@@ -39,11 +43,14 @@
     "cordova-plugin-ionic-webview": "^1.1.16",
     "cordova-plugin-splashscreen": "^4.0.3",
     "cordova-plugin-whitelist": "^1.3.1",
+    "date-fns": "^1.29.0",
     "immutable": "^3.8.2",
     "ionic-angular": "3.9.2",
     "ionic-plugin-keyboard": "^2.2.1",
     "ionicons": "3.0.0",
-    "rxjs": "5.5.2",
+    "ngx-pagination": "^3.2.1",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "sockjs-client": "^1.1.5",
     "sw-toolbox": "3.6.0",
     "websocket": "^1.0.26",
