@@ -1,84 +1,24 @@
-import 'rxjs/add/operator/toPromise';
+import { Injectable } from "@angular/core";
+import { Rest } from "../rest/rest";
+import { map, catchError } from "rxjs/operators";
+import { of as observableOf, Observable } from "rxjs";
+import { HttpErrorResponse } from "@angular/common/http";
+import { User } from "../../models/user";
 
-import { Injectable } from '@angular/core';
-
-import { Api } from '../api/api';
-
-/**
- * Most apps have the concept of a User. This is a simple provider
- * with stubs for login/signup/etc.
- *
- * This User provider makes calls to our API at the `login` and `signup` endpoints.
- *
- * By default, it expects `login` and `signup` to return a JSON object of the shape:
- *
- * ```json
- * {
- *   status: 'success',
- *   user: {
- *     // User fields your app needs, like "id", "name", "email", etc.
- *   }
- * }Ø
- * ```
- *
- * If the `status` field is not `success`, then an error is detected and returned.
- */
 @Injectable()
-export class User {
-  _user: any;
+export class UserProvider {
+  constructor(private rest: Rest) {}
 
-  constructor(public api: Api) { }
-
-  /**
-   * Send a POST request to our login endpoint with the data
-   * the user entered on the form.
-   */
-  login(accountInfo: any) {
-    let seq = this.api.post('login', accountInfo).share();
-
-    seq.subscribe((res: any) => {
-      // If the API returned a successful response, mark the user as logged in
-      if (res.status == 'success') {
-        this._loggedIn(res);
-      } else {
-      }
-    }, err => {
-      console.error('ERROR', err);
-    });
-
-    return seq;
-  }
-
-  /**
-   * Send a POST request to our signup endpoint with the data
-   * the user entered on the form.
-   */
-  signup(accountInfo: any) {
-    let seq = this.api.post('signup', accountInfo).share();
-
-    seq.subscribe((res: any) => {
-      // If the API returned a successful response, mark the user as logged in
-      if (res.status == 'success') {
-        this._loggedIn(res);
-      }
-    }, err => {
-      console.error('ERROR', err);
-    });
-
-    return seq;
-  }
-
-  /**
-   * Log the user out, which forgets the session
-   */
-  logout() {
-    this._user = null;
-  }
-
-  /**
-   * Process a login/signup response to store user data
-   */
-  _loggedIn(resp) {
-    this._user = resp.user;
+  public fetchUser(): Observable<User> {
+    return this.rest.getUserInfo().pipe(
+      map(
+        ({ code, email, message, ...rest }) =>
+          ({
+            ...rest,
+            login: email
+          } as any)
+      ),
+      catchError((err: HttpErrorResponse) => observableOf(err))
+    );
   }
 }
