@@ -1,3 +1,4 @@
+import { map, take, tap } from "rxjs/operators";
 import { Component, ChangeDetectionStrategy } from "@angular/core";
 import { Observable } from "rxjs";
 import { Pending } from "../../core/models/pending";
@@ -6,6 +7,7 @@ import { AppState } from "../../core/app-state";
 import * as fromActions from "../../core/actions/pending.actions";
 import { Rest } from "../../core/providers/rest/rest";
 import { PendingProvider } from "../../core/providers/pending-provider/pending-provider";
+import format from 'date-fns/format'
 
 @Component({
   selector: "pending-container",
@@ -13,24 +15,29 @@ import { PendingProvider } from "../../core/providers/pending-provider/pending-p
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class PendingContainerComponent {
-  pendings$: Observable<Pending[]>;
+  public pendings$: Observable<Pending[]>;
+  public pendingsMarked$: Observable<boolean>;
+  public date: string = format(new Date(), 'DD-MM-YYYY');
 
   constructor(private store: Store<AppState>, private prov: PendingProvider) {
     this.pendings$ = this.store.select((state) => state.pending.data);
+    this.pendingsMarked$ = this.pendings$.pipe(
+      map((pendings: Pending[]) =>
+        pendings.filter((pending: Pending) => pending.isMarked)
+      ),
+      map((marked: Pending[]) => marked.length > 0)
+    );
   }
 
-  create(): void {
-    this.prov.createPendings(1)
-      .subscribe(() => this.store.dispatch(new fromActions.FetchPendings({})))
-  }
-
-  complete(): void {
+  public complete(): void {
     this.pendings$
-      .take(1)
-      .map((pendings: Pending[]) =>
-        pendings
-          .filter((pending: Pending) => pending.isMarked)
-          .map((pending: Pending) => pending.id)
+      .pipe(
+        take(1),
+        map((pendings: Pending[]) =>
+          pendings
+            .filter((pending: Pending) => pending.isMarked)
+            .map((pending: Pending) => pending.id)
+        )
       )
       .subscribe((ids: string[]) => {
         if (ids.length !== 0) {
@@ -39,13 +46,15 @@ export class PendingContainerComponent {
       });
   }
 
-  cancel(): void {
+  public cancel(): void {
     this.pendings$
-      .take(1)
-      .map((pendings: Pending[]) =>
-        pendings
-          .filter((pending: Pending) => pending.isMarked)
-          .map((pending: Pending) => pending.id)
+      .pipe(
+        take(1),
+        map((pendings: Pending[]) =>
+          pendings
+            .filter((pending: Pending) => pending.isMarked)
+            .map((pending: Pending) => pending.id)
+        )
       )
       .subscribe((ids: string[]) => {
         if (ids.length !== 0) {
