@@ -1,3 +1,4 @@
+import { take, pluck, tap, switchMap, map } from "rxjs/operators";
 import { Injectable } from "@angular/core";
 import { Actions, Effect } from "@ngrx/effects";
 import { HttpErrorResponse } from "@angular/common/http";
@@ -7,7 +8,7 @@ import * as fromUserActions from "../actions/user.actions";
 import * as fromLoaderActions from "../actions/_loader.actions";
 import * as fromToastActions from "../actions/_toast.actions";
 import * as fromRouteActions from "../actions/_route.actions";
-import { LoginProvider, Settings } from "../providers/providers";
+import { LoginProvider } from "../providers/providers";
 import { LoginFormState } from "../reducers/login-form/login-form.reducer";
 import { Store } from "@ngrx/store";
 import { AppState } from "../app-state";
@@ -24,31 +25,33 @@ export class LoginFormEffects {
   ) {}
 
   @Effect()
-  public loginUser$ = this.actions$
-    .ofType(fromActions.SUBMIT_FORM)
-    .pluck("payload")
-    .do((payload: LoginFormState) => this.handleSideBefore())
-    .switchMap((payload: LoginFormState) =>
+  public loginUser$ = this.actions$.ofType(fromActions.SUBMIT_FORM).pipe(
+    pluck("payload"),
+    tap(() => this.handleSideBefore()),
+    switchMap((payload: LoginFormState) =>
       this.loginProvider.loginUser(payload.data)
-    )
-    .do((res: null | HttpErrorResponse) => this.handleSideAfter(res))
-    .map((res: null | HttpErrorResponse) => this.handleLoginRes(res));
+    ),
+    tap((res: null | HttpErrorResponse) => this.handleSideAfter(res)),
+    map((res: null | HttpErrorResponse) => this.handleLoginRes(res))
+  );
 
-  handleLoginRes(res: null | HttpErrorResponse): fromActions.LoginFormActions {
+  private handleLoginRes(
+    res: null | HttpErrorResponse
+  ): fromActions.LoginFormActions {
     return res instanceof HttpErrorResponse
       ? new fromActions.SubmitFormFail()
       : new fromActions.SubmitFormSucc();
   }
 
-  handleSideBefore(): void {
+  private handleSideBefore(): void {
     this.store.dispatch(new fromLoaderActions.Show("Logowanie..."));
   }
 
-  handleSideAfter(res: null | HttpErrorResponse): void {
+  private handleSideAfter(res: null | HttpErrorResponse): void {
     this.store.dispatch(new fromLoaderActions.Hide());
     if (!(res instanceof HttpErrorResponse)) {
       this.updateUser();
-      this.store.dispatch(new fromRouteActions.Root(new _Route("dashboard")));
+      this.store.dispatch(new fromRouteActions.Push(new _Route("dashboard")));
     } else {
       this.store.dispatch(
         new fromToastActions.Show(
@@ -58,13 +61,15 @@ export class LoginFormEffects {
     }
   }
 
-  updateUser(): void {
+  private updateUser(): void {
     this.store
       .select((state) => state.loginForm)
-      .take(1)
-      .do((res: LoginFormState) => sessionStorage.__mail = res.data.login)
+      .pipe(
+        take(1),
+        tap((res: LoginFormState) => (sessionStorage.__mail = res.data.login))
+      )
       .subscribe((data: LoginFormState) => {
-        this.store.dispatch(new fromUserActions.UpdateUser(data));
+        this.store.dispatch(new fromUserActions.FetchUser());
       });
   }
 }
