@@ -1,13 +1,16 @@
+import { pluck, tap, map } from "rxjs/operators";
 import { Injectable } from "@angular/core";
 import { Actions, Effect } from "@ngrx/effects";
 
 import * as fromActions from "../actions/offer-form.actions";
 import * as fromOfferActions from "../actions/offer.actions";
+import * as fromToastActions from "../actions/_toast.actions";
 import { Store } from "@ngrx/store";
 import { AppState } from "../app-state";
 import { OfferFormState } from "../reducers/offer-form/offer-form.reducer";
 import { OfferProvider } from "../providers/providers";
 import { Offer } from "../models/offer";
+import { HttpErrorResponse } from "@angular/common/http";
 
 @Injectable()
 export class OfferFormEffects {
@@ -18,22 +21,33 @@ export class OfferFormEffects {
   ) {}
 
   @Effect()
-  public updateForm = this.actions$
-    .ofType(fromActions.UPDATE_FORM)
-    .pluck("payload")
-    .map(([prevState, currState]) => this.offerProvider.computeOffer(prevState, currState))
-    .map((offer: Offer) => Object.assign({
-      data: offer
-    }))
-    .map((payload: OfferFormState) => new fromActions.UpdateFormSucc(payload));
-
-  @Effect()
-  public submitForm$ = this.actions$
-    .ofType(fromActions.SUBMIT_FORM)
-    .pluck("payload")
-    .do((payload: OfferFormState) =>
-      this.store.dispatch(new fromOfferActions.UpdateOffer(payload.data))
+  public submitForm$ = this.actions$.ofType(fromActions.SUBMIT_FORM).pipe(
+    pluck("payload"),
+    map(this.checkForFields),
+    tap((payload: OfferFormState | HttpErrorResponse) =>
+      this.store.dispatch(
+        payload instanceof HttpErrorResponse
+          ? new fromToastActions.Show("Proszę, uzupełnij wszystkie pola.")
+          : new fromOfferActions.UpdateOffer(payload.data)
+      )
+    ),
+    map(
+      (payload: OfferFormState | HttpErrorResponse) =>
+        payload instanceof HttpErrorResponse
+          ? new fromActions.SubmitFormFail()
+          : new fromActions.SubmitFormSucc(payload)
     )
-    .map((payload: OfferFormState) => new fromActions.SubmitFormSucc(payload));
+  );
 
+  private checkForFields(
+    payload: OfferFormState
+  ): OfferFormState | HttpErrorResponse {
+    let { name, price, discount, count, startDate, endDate } = payload.data;
+
+    return name && price && discount && count && startDate && endDate
+      ? payload
+      : new HttpErrorResponse({
+          status: -1
+        });
+  }
 }
