@@ -1,24 +1,25 @@
 export class Offer {
   constructor(offer?: Offer) {
-    if (offer) {
-      Object.assign(this, offer);
-      if (!offer.calculatedPrice) {
-        this.calculatedPrice = calculatePrice(this.price, this.discount);
-      }
-    } 
+    Object.assign(this, offer);
   }
 
-  public id: string | number = '';
+  public id?: string | number = '';
   public name: string = '';
-  public restaurantId: string | number = '';
+  public restaurantId?: string | number = '';
   public image: string = null;
   public description?: string = '';
   public price?: number = null;
   public calculatedPrice?: number = null;
+  public discountMetric?: DiscountMetric = DiscountMetric.PERCENTAGE;
   public discount?: number = null;
   public count?: number = null;
+  public receiveTimeStart?: number;
+  public receiveTimeEnd?: number;
+  public startDate?: string = "";
+  public endDate?: string = "";
 }
 
-const calculatePrice = (price, discount) => {
-  return Math.round(price*(100 - discount)/10)/10;
+export enum DiscountMetric {
+  PERCENTAGE = "PERCENTAGE",
+  PRICE =  "PRICE"
 }
