@@ -1,10 +1,14 @@
 import { LoginForm } from "./login-form";
 
 export class User extends LoginForm {
-  
   constructor() {
     super();
   }
+
+  public openHours?: OpenSchedule;
+  public website: string;
+  public description: string;
+  public image: string;
   
 }
 
@@ -13,3 +17,14 @@ export interface User extends LoginForm {
   description: string,
   image: string
 };
+
+export interface OpenSchedule {
+  week?: OpenHours,
+  saturday?: OpenHours,
+  sunday?: OpenHours
+}
+
+export interface OpenHours {
+  openTime: string,
+  closeTime: string
+}
