@@ -1,9 +1,9 @@
 import { Api } from "./api/api";
 import { Rest } from "./rest/rest";
 import { Settings } from "./settings/settings";
-import { User } from "./user/user";
+import { UserProvider } from "./user/user";
 import { OfferProvider } from "./offer-provider/offer-provider";
 import { LoginProvider } from "./login-provider/login-provider";
 import { StompProvider } from "./stomp/stomp-provider";
 
-export { Api, Rest, Settings, User, OfferProvider, LoginProvider, StompProvider };
+export { Api, Rest, Settings, UserProvider, OfferProvider, LoginProvider, StompProvider };
